스토어 속성 모듈화
