+++ projects/vue-todo/src/store/getters.js
@@ -1,0 +1,3 @@
+export const storedTodoItems = (state) => {
+    return state.todoItems;
+}
